% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.mmdt.R
\name{fig.mmdt}
\alias{fig.mmdt}
\title{Plot Results of Multi-Modal Density Test}
\usage{
fig.mmdt(mmdt.results, output = "p-value", coords = c(NA, NA))
}
\arguments{
\item{mmdt.results}{an object resulting from the 'mmdt' command.}

\item{output}{type of image to be produced. Can be "t-statistic" or "p-value". Default is "p-value".}

\item{coords}{a vector of length d [e.g., c(NA, NA, 3.25) for d=3] giving the coordinates at which the plane should be visualized.
Only necessary for plotting results that include three or more dimensions.
Entries should be "NA" for the two modalities to be plotted, and other entries should give the
value along the other dimensions at which the results should be plotted.}
}
\value{
A description of overall group differences, and approximate regions that drive overall differences
}
\description{
This function creates a summary of the mmdt output.
}
\examples{
\dontrun{
mmdt.obj = get.mmdt.obj(masks = masks, modal1 = t1s, modal2 = flairs,
                        ids = ids, groups = groups)
results = mmdt(mmdt.obj)

fig.mmdt(results, output="p-value")}
}
