% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.mmdt.obj.R
\name{get.mmdt.obj}
\alias{get.mmdt.obj}
\title{Get Multi-Modal Density Test Object}
\usage{
get.mmdt.obj(masks, modal1, modal2, modal3 = NULL, modal4 = NULL,
  modal5 = NULL, modal6 = NULL, ids, groups, parallel = TRUE,
  cores = 2, pb = TRUE)
}
\arguments{
\item{masks}{a vector of class \code{character} that gives .nii or .nii.gz filenames for subjects' masks.
Masks will demarcate which voxels will be included in the analysis, and should be coded by TRUE/FALSE or 1/0.}

\item{modal1}{vector of class \code{character} that give .nii or .nii.gz filenames for a given imaging modality across subjects.
At least two modalities (modal1 and modal2) must be entered. Up to 6 can be included.}

\item{modal2}{see 'modal1' description}

\item{modal3}{see 'modal1' description}

\item{modal4}{see 'modal1' description}

\item{modal5}{see 'modal1' description}

\item{modal6}{see 'modal1' description}

\item{ids}{a vector of subject ids. Must be the same length as the filenames in the 'modal#' vectors.}

\item{groups}{a vector of group membership. Must be two categories, and should be the same length as 'ids'.}

\item{parallel}{is a logical value that indicates whether the user's computer
is Linux or Unix (i.e. macOS), and should run the code in parallel.}

\item{cores}{if parallel = TRUE, cores is an integer value that indicates how many cores
the function should be run on.}

\item{pb}{is a logical value that indicates whether or not a progress bar will be shown during analysis.}
}
\value{
An mmdt object to be used in the 'mmdt' function.
}
\description{
This function creates an mmdt object from lists of \code{nifti} filenames.
}
\examples{
\dontrun{
masks = c("mask01.nii", "mask02.nii", "mask03.nii", "mask04.nii")
t1s = c("t101.nii", "t102.nii", "t103.nii", "t104.nii")
flairs = c("flair01.nii", "flair02.nii", "flair03.nii", "flair04.nii")
ids = c(1, 2, 3, 4)
groups = c(1, 1, 2, 2)

mmdt.obj = get.mmdt.obj(masks = masks, modal1 = t1s, modal2 = flairs,
                        ids = ids, groups = groups)}
}
