% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmdt.R
\name{mmdt}
\alias{mmdt}
\title{Multi-Modal Density Test}
\usage{
mmdt(mmdt.obj, mins = NULL, maxs = NULL, gridsize = NULL, H = NULL,
  mc.adjust = "BY", nperm = 500, parallel = TRUE, cores = 2,
  pb = TRUE)
}
\arguments{
\item{mmdt.obj}{an mmdt object obtained using the get.mmdt.obj function.}

\item{mins}{a vector giving the lower intensity bounds for each modality.
If NULL, lower bounds will be set to the minimum observed value for each modality.
Length of this vector should be equal to the number of modalities being analyzed.}

\item{maxs}{a vector giving the upper intensity bounds for each modality.
If NULL, upper bounds will be set to the maximum observed value for each modality.
Length of this vector should be equal to the number of modalities being analyzed.}

\item{gridsize}{is a vector giving the number of points along each dimension at which the densities should be evaluated and tested.
If NULL, this value defaults to 151x151 for two modalities, 51x51x51 for three, and 21x21x21x21 for four.
Must be specified manually when analyzing 4-6 modalities.}

\item{H}{is the bandwidth matrix used for kernel density estimation. If NULL, a plug-in bandwidth estimator is used.}

\item{mc.adjust}{is a string defining the method(s) of multiple comparisons to use.
Default is "BY", which controls FDR using the Benjamini-Yekutieli procedure.
The additional options are: "maxt", which controls global FWER using max-t correction,
and "tfce", which controls global FWER using threshold-free cluster enhancement.
Both of the latter options use permutation to determine significance.}

\item{nperm}{is an integer value that gives the number of permutations desired.}

\item{parallel}{is a logical value that indicates whether the user's computer
is Linux or Unix (i.e. macOS), and should run the code in parallel.}

\item{cores}{if parallel = TRUE, cores is an integer value that indicates how many cores
the function should be run on.}

\item{pb}{is a logical value that indicates whether or not a progress bar will be shown during analysis.}
}
\value{
A list containing subject.densities (kernel density estimates for each subject),
teststat.matrix (a matrix giving test statistics at each evaluated point),
pval.matrix.uncorrected (a matrix giving uncorrected p-values at each evaluated point),
pval.matrix.perm (a matrix giving permutation-corrected p-values at each evaluated point),
evaluated.points (a matrix giving the intensity values at which the test was evaluated for each modality).
}
\description{
This function runs the multi-modal density test (mmdt) using an mmdt object obtained from get.mmdt.obj
}
\examples{
\dontrun{
mmdt.obj = get.mmdt.obj(masks = masks, modal1 = t1s, modal2 = flairs,
                        ids = ids, groups = groups)
results = mmdt(mmdt.obj)

eval = results$evaluated.points
t.mat = results$teststat.matrix
p.mat = results$pval.matrix.perm}
}
