% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.mmdt.R
\name{summarize.mmdt}
\alias{summarize.mmdt}
\title{Get Multi-Modal Density Test Summary}
\usage{
summarize.mmdt(mmdt.results)
}
\arguments{
\item{mmdt.results}{an object resulting from the 'mmdt' command.}
}
\value{
A description of overall group differences, and approximate regions that drive overall differences
}
\description{
This function creates a summary of the mmdt output.
}
\examples{
\dontrun{
mmdt.obj = get.mmdt.obj(masks = masks, modal1 = t1s, modal2 = flairs,
                        ids = ids, groups = groups)
results = mmdt(mmdt.obj)

summarize.mmdt(results)}
}
